+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Adds dynamic entry file support
+
 ## 10.13.0 (2022-03-17)
 
 * Adds Instabug Experiments APIs